Lesson15 9.投稿機能の課題

+++ app/Http/Controllers/TasksController.php
@@ -11,11 +11,22 @@ class TasksController extends Controller
     // get で tasks/ にアクセスされた場合の「一覧表示処理」
     public function index()
     {
-        $tasks = Task::all();
+        // $tasks = Task::all();
+        // return view('tasks.index', [
+        //     'tasks' => $tasks,
+        // ]);
+        $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
         
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+        return view('tasks.index', $data);
     }
 
     // get で tasks/create/ にアクセスされた場合の「新規登録画面表示処理」
@@ -36,10 +47,16 @@ class TasksController extends Controller
             'content' => 'required',
         ]);
         
-        $task = new Task;
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
+        // $task = new Task;
+        // $task->content = $request->content;
+        // $task->status = $request->status;
+        // $task->save();
+        // return redirect('/');
+        
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+        ]);
         
         return redirect('/');
     }
@@ -49,9 +66,13 @@ class TasksController extends Controller
     {
         $task = Task::find($id);
         
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.show', [
+                'task' => $task,
+            ]);
+        } else {
+            return redirect('/');
+        }
     }
 
     // get で tasks/id/edit にアクセスされた場合の「更新画面表示処理」
@@ -59,9 +80,13 @@ class TasksController extends Controller
     {
         $task = Task::find($id);
         
-        return view('tasks.edit', [
-            'task' => $task,
-        ]);
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.edit', [
+                'task' => $task,
+            ]);
+        } else {
+            return redirect('/');
+        }
     }
 
     // put または patch で tasks/id にアクセスされた場合の「更新処理」
@@ -83,8 +108,13 @@ class TasksController extends Controller
     // delete で tasks/id にアクセスされた場合の「削除処理」
     public function destroy($id)
     {
-        $task = Task::find($id);
-        $task->delete();
+        // $task = Task::find($id);
+        // $task->delete();
+        
+        $task = \App\Task::find($id);
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
         
         return redirect('/');
     }
